+++ CHANGELOG.md
@@ -6,8 +6,9 @@ All notable changes to this project will be documented in this file.
 
 - Add `COMMAND INFO` command (ACL categories, tips, key specifications and subcommands are not yet supported)
 - Add `COMMAND HELP` command
-- Support no arguments for `COMMAND` command
 - Add `SELECT` command
+- Support no arguments for `COMMAND` command
+- Support authentication for `HELLO` command
 
 ## [v0.1.0] - 2023-02-26